Update patch set 1
Patch Set 1:

Stuart also have the same idea that we do not need wide version of this instruction.

What I was thinking is: For all other instructions which take RegR(X or W) arguments, we provide wide version for RegX and unwide version for RegW. If people read assembly syntax(TBZ <R><t>, #<imm>, <label>) and do not care about how the instruction is encoded. They will feel natural that we share the same rule on this instruction also. This might not match the encoding format very well. But I think it is more natural to define in this way. The rest things(opcode, register, imm) will be fixed-up in the assembler.

But it seems that I am the minority. :)

Patch-set: 1
Reviewer: Zheng Xu <1042649@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0