Update patch set 1
Patch Set 1:

This is a mips-specific file. The issue here was that the negative value was preferred, but the branch-over happened on the positive condition. So you got the exact inverse values.

I am not sure what values are optimal overall. I was just curious, as there is a branch incurred. I don't know enough about current Mips implementations to know whether a forward branch has significant penalties (like misprediction, if Mips does speculation), etc. Maybe the optimal thing is a conditional move. That's why I asked.

If you're answer is just: "it doesn't matter too much," that's OK with me and I'll merge.

Patch-set: 1