Update patch set 1
Patch Set 1:

The instruction sequence can be optimized on MIPS by using
xor and sltiu instructions instead of using branch. The instruction sltiu will need to be added to the MIPS assembler for that.

In the mean time, I would appreciate if this patch can be committed as this is definite bug fix which helps move the MIPS32 boot a lot further.

Patch-set: 1