Update patch set 1
Patch Set 1:

It seems like the app in question is hitting the max allocation stack size before triggering a concurrent GC. What is this situation like. A big heap and a lot of small objects?

As it is, I'd be worried about the consequence on allocation performance since it adds some code to the allocation fast path, which we'd like to avoid as much as possible.

In this situation, my first inclination is, instead, to bump up the max allocation stack size. For example, up to 2 MB. With that, would the problem practically go away? Note that when the extra stack space isn't used, it won't cost PSS/RSS.

Patch-set: 1
Label: Code-Review=0