Update patch set 1
Patch Set 1:

I'm kind of neutral on this change. The main thing it does is avoid crashes dumping SSA representation for the phi opcode. Why would we have phi opcodes and not have SSA? The changes to branch make something an "if" rather than an "else if" and just seem to be moving things around. If you could demonstrate the problem to us then it'd be easier for me to merge. Thanks!

Patch-set: 1