Update patch set 1
Patch Set 1:

As I understand SSA form is build only for reachable BBs. if BB becomes unreachable or MIR is created by compiler and it is not yet updated with SSA form. Other case will be to dump CFG before SSA built (for this case patch affects not only Phi but branch MIRs as well).
So yes, the general idea is to make GetDalvikDissasembly to be immutable to SSA generation. This is mostly for debugging.

Patch-set: 1