Update patch set 1
Patch Set 1:

Guys, when look deeper in this case, I guess we need a more general solution to handle "diamonds" in the backend while lowering one MIR. We need to copy the Register Allocator state in the beginning, reset it before each new path and merge at the end. The API could look as follows: some class which copies RegInfos in constructor, each time we are resetting state we should remember which physical regs are corrupted. When we merge all patches we should clobber all regs which are corrupted in any of these paths...

Razvan's please note that the fix of this specific issue is already merged to AOSP: https://android-review.googlesource.com/#/c/102980/. But this CL is better because that one blocks TrackLiveTemps backend optimization for this particular case. I would suggest to revert it in this patch.

Opinions?

Patch-set: 1
Reviewer: Serguei I Katkov <1040038@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0