Update patch set 3
Patch Set 3: Code-Review-2

(7 comments)

Thanks Daniel for bringing this up. It's interesting how you optimize single stepping. That's not exactly what I had in mind though but that opens the discussion. The point to use full deoptimization to handle single-step was to make it functional first. But it's true we can do better job in this area, especially for intensive computational code.

My main concern with this CL is the overall complexity of optimizing single-step. I'm particularly concerned of all the new stuff in Dbg::UpdateDebugger. This is the hottest method when debugging since it is used to report JDWP location events for each instruction. Besides the code to deal with method call preparation may have an impact of interpreter itself (without debugging). Please see my comments.

Nevertheless, I find your point about STEP OVER interesting. When stepping over a method call, it's true we shouldn't need to deoptimize. If we need to report other JDWP location events (breakpoint, method entry/exit and field access/modification), we would already have deoptimize (either the method with a breakpoint of everything for the other events) so we would automatically execute with interpreter in these cases and do detect them.

The raw plan I had in mind (but never experienced yet) to optimize single-stepping is:
1) STEP INTO: force the use of interpreter entrypoint. When invoking a method from the interpreter, we could force the use of interpreter entrypoint. No need to deoptimize the invoked method (ie update its entrypoint). The advantage is only the "event" thread would execute with interpreter, not the other ones. We'd still need to deoptimize "compiled" caller when exiting an interpreted method with that kind of step.
2) STEP OUT: trigger a stack deoptimization so we can continue execution of compiled code with interpreter.

Patch-set: 3
Label: Code-Review=-2