Update patch set 3
Patch Set 3:

Dear Ian, Dear Sebastien, Thank you for taking the time to review our patch.  We will try to refactor the patch so there is minimal overhead in the UpdateDebugger() method. Regarding the new instrumentation events: We introduced them to provide clean separation between the debugger and the rest of the system. Do you have a better suggestion? Should we call the methods of the debugger directly from the interpreter for example?

Patch-set: 3