Update patch set 4
Patch Set 4:

This is a work in progress patch. It implements the STEP_IN idea suggested by Sebastien. Could you please take a look so we know that we are going in the right direction? We will submit the stackwalk optimization patch separately, because we are still working on a solution to eliminate the change in the ShadowFrame layout. I will upload a new patchset tomorrow, rebased to the latest source.

Patch-set: 4
Label: Verified=0