Update patch set 2
Patch Set 2:

If we have a read-barrier based GC you often have to visit C heap roots at safe points, such as with C4. If the GC were to transition during a Visit then it would cause a large number of root visits and create an outlier for pause time.

Patch-set: 2