Update patch set 3
Patch Set 3:

> The SP and PC can be in the list in A32 (ARM) instructions, but not
 > in T32 (Thumb) instructions. Also the usage of instructions where
 > the base register is part of the list is not recommended. Please
 > see section: F7.1.212 STM (STMIA, STMEA) in ARM ARMv8.
 > 
 > http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.genc010197a/index.html

Thanks for review. But now, it is impossible to use r14. I think Encoding Map has to be defined likes Thumb2Stmia.
Encoding Map is not good position to disallow.

 > The SP and PC can be in the list in A32 (ARM) instructions, but not
 > in T32 (Thumb) instructions. Also the usage of instructions where
 > the base register is part of the list is not recommended. Please
 > see section: F7.1.212 STM (STMIA, STMEA) in ARM ARMv8.
 > 
 > http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.genc010197a/index.html

Thanks for review.
Now, we can not use r14. It can be used with stm. 
I think EncodingMap is not good location to disallow SP, PC register.
Because there is no way to use r14 which is can be used.
Serban, how do you think about r14?

Patch-set: 3
Label: Verified=0