Update patch set 5
Patch Set 5:

(7 comments)

I think that's going in the right direction (at least the one I had in mind). Do that work with method invoked through reflection or from JNI? 

Once it's functional, we'll have to evaluate the performance impact. While it's probably a net win for debugging, I don't want it to slow down the interpreter performance when not debugging.

Patch-set: 5