Update patch set 3
Patch Set 3: Code-Review-2

The encode all is still needed here.   We need to turn the implicit null-check loads into scheduling barriers.  Consider the following sequence:

  copy r3, r4
  load r1, (r2)    ; load used for implicit null check
SafePointPcLabel:

The above load starts off with r1 def, r2 and memory use.  In the current code, we'd set all def/use bits on the load, turning it into a scheduling barrier.  The recovery information would be associated with SafePointPcLabel (which is where the null check handler would look should the load fault).

However, with your proposed change, it would be legal for load hoisting to rearrange the code as follows:

  load r1, (r2)
  copy r3, r4
SafePointPcLabel:

In this case, if we faulted on the load the null check handler would expect to find recovery information associated with the next address.  However, it would not - the information would still be associated with SafePointPcLabel:.

Let me know if I've misunderstood your intention here.

Patch-set: 3
Label: Code-Review=-2