Update patch set 16
Patch Set 16:

You should avoid using dex2oatd for timing, just use dex2oat (the checking overhead is quite large).  I also find it useful to time with -j1 and -j4 (or something we'd expect on an actual device).  Finally, it is also useful to use the --dump-timing option to dex2oat (though you'd not want to report total time when that one is on - again, it adds overhead).

Patch-set: 16
Label: Verified=0