Update patch set 1
Patch Set 1:

> > You code looks fine by me, but I'd like to have a test with it.
 > > Adding a test that forces materialization should be easy: just
 > > construct a graph with the condition of an if not being just
 > before
 > > the if.
 > 
 > Ok I'll do that.

I'm working on this. I think I won't be able to achieve this by using the DEX bytecode, as the HIf and HCondition are generated as a pair by the HGraphBuilder (in HGraphBuilder::If_22t()). So instead I am trying to manually creating a HGraph, compiling it and running the code. Was it what you had in mind?

Patch-set: 1