Update patch set 18
Patch Set 18: Code-Review+1

Bill, this is functionality that we are interested in having :) Would you like me to run some compile time numbers to show the overhead from this change? I expect it won't change much since the only real change is caching of an additional field.

Patch-set: 18
Label: Code-Review=+1