Update patch set 4
Patch Set 4:

I have a few other comments on this but I'm still working on following everything.

Until then, could you run some benchmarks and see if this improves the performance in any relevant way? The CL does add a bit of complexity and if the improvement is not much it might worth simplifying stuff.

Patch-set: 4