Update patch set 2
Patch Set 2:

> General comment, getting rid of uses of WaitHoldingLocks is awesome! What remains blocking just deleting it? Perhaps we can protect it and make ThreadPool a friend class..

It’s still used by ThreadPool to support the parallel workers in MarkSweep. The pattern is that when the main GC thread waits for the worker threads to finish work, the main thread uses WaitHoldingLocks while holding the exclusive heap bitmap lock and the shared mutator lock if a concurrent phase or the exclusive mutator lock if at a pause.

It seems it’s not easy to not hold those locks while waiting because 1) the main thread can’t unlock the mutator lock in the middle of a pause, and 2) the heap bitmap is written by multiple worker threads, which can’t be expressed using the reader-writer lock (so, the main thread exclusively holds the heap bitmap lock on behalf.)

I protected it and made ThreadPool a friend class.

Patch-set: 2