Update patch set 1
Patch Set 1:

This means that we are calling the system assembler, which is from gas/binutils, and not necessarily gcc (although we may be calling through a gcc wrapper; I would have to check). The alternative here is to stop using non-standard features of the assembler. It is hard for me to believe that these .altmacro/LOCAL's are providing such a large benefit in readability for assembly that it can't be rewritten to use something equally clear. Even if you wrote a custom generator for your assembly files (because you might have a huge number of these), I think it would be preferable to depending on gas-specific features (especially when the resulting code is still hard to read).

Patch-set: 1