Update patch set 1
Patch Set 1:

The proposed solution here looks a little fragile.  It seems to me that the issue here is that a 32-bit view of a 64-bit register is being blindly used as 64 bits.  Rather than a directed workaround for this particular incarnation of the problem, I wonder if it may be cleaner to go for something a bit more general.

I haven't though this through completely, but I'd explore something like this:

NarrowRegLoc() simply changes the view - implicitly discarding the upper 32 bits.  The result isn't intended to be widened later.  But, suppose a flag were added to RegisterInfo that NarrowRegLoc() (or any other similar operation which leaves garbage in the high bits) could set.  Then, the utility As64Bit() - which right now just DCHECKS to ensure that there is a corresponding 64-bit register view for the given 32-bit register view - would test this bit and actually generate an extend operation if necessary.

Doing something like this would avoid conservatively generating useless extends - but would still ensure you got them where needed.  Also, it might be applicable to other targets.  You'd need to be diligent about using As64Bit() [or perhaps a similar utility that returned the natural word size view depending on x86 or x86_64] - but you probably should be doing that anyway.

Patch-set: 1