Update patch set 1
Patch Set 1:

(1 comment)

Bill, I thought about similar approach but it has performance disadvantage. However I think I can combine your proposal and my solution. First what is the problem with your proposal. This is use case:
promoted register is used for narrowing down, we mark it is having garbage in high part. Now in two different path of one branch it is used as index in array access. It means we will need to put 32-bit to 64-bit in both conversion (it is only one redundant instruction) however this case shows that we cannot clear the garbage flag as soon as we insert the conversion instruction and that leads to other use case. long-to-int and aget/aput pair with the same index and in the same branch. As a result we will insert redundant conversion for iput. And it is happens in a loop it is undesirable. So the possible solution would be still generation nopped 32-bit to 64-bit conversion and used this LIR as a garbage flag. In this case As64Bit() can just remove nop in case LIR is installed. However the efficiency will require us to remove installed LIR each time the physical register is re-assigned. I need to check how easy to implement it (I mean to catch all paths leading to assignment to this physical register). What do you think?

Patch-set: 1