Update patch set 1
Patch Set 1:

I think the problem is really what convention we use:

1. If we demand that registers associated to 32-bit integers never have "garbage" on the most significant 32-bits then we'll have to do the zero-extend at the source (i.e. where the value is originated).

2. If, on the other hand, we allow "garbage", then we have to do the zero-extend at the destination (i.e. where the value is consumed). This boils down to always access 32-bit integers via 32-bit views of the registers where they are contained (which - if possible - is probably a good idea anyway).

For Arm64 it is quite possible to do the second, as instructions typically allow to specify views for registers. For example, it is possible to sign extend the register used as an index in a load/store, without requiring an extra separate instruction (e.g. `ldr x0, [sp, w2, sxtw]'). I believe that - if this is done consistently - then the current implementation of long-to-int can work. The sign extension can be done in LoadBaseIndexed which - for arm64 - does not require additional instructions.

For x86, however, doing the sign extension in LoadBaseIndexed() may require a separate zero-extend instruction, which would then become redundant in the cases where long-to-int has two distinct operands. In the latter case, indeed, the mov instruction generated in long-to-int will do the zero-extend for free.

However, I agree with Bill when he says that handling the `long-to-int vN, vN' case with an explicit conversion shouldn't have a noticeable impact on performance. Typically long-to-int-s have distinct arguments. It is then probably a good idea to just do it this way.

PS: I tried doing what Bill suggested on Arm64,

  long-to-int v2, v2
  return v2

I had to modify the dex file by hand to achieve this (the compiler generated a `long-to-int v0, v2; return v0'). I could not, however, trigger the bug. I tried other various things and switched promotion on and off, but still I couldn't trigger the bug. This happened because the register always ended up being used via a 32-bit view.

Patch-set: 1