Update patch set 4
Patch Set 4:

The fact that it is hitting a stack overflow error is a problem.  This should not happen in a real app.  I think the app needs to pass a bigger stack to its threads.  We've seen some apps using a 32K stack (and maybe smaller) in a misguided attempt to save memory.

Patch-set: 4