Update patch set 1
Patch Set 1:

Many thanks, Serguei. I could reproduce the bug via the following Java code:

  public class Main {
    private static int Fn1(int[] out_array) {
        long idx = 0xffffff00000000l;
        return out_array[(int)idx];
    }
    public static void main(String[] args) {
        int[] a = new int[10];
        a[0] = 1234;
        System.out.printf("%d\n", Fn1(a));
    }
  }

Which triggers the bug also on Arm64. As far as I can see, however, the problem for Arm64 is in int_arm64.cc, at the end of the GenArrayGet implementation, where As64BitReg() is used to force a 64-bit view of the index:

    if (rl_result.ref) {
      LoadRefIndexed(reg_ptr, As64BitReg(rl_index.reg), rl_result.reg, scale);
    } else {
      LoadBaseIndexed(reg_ptr, As64BitReg(rl_index.reg), rl_result.reg, scale, size);
    }

The problem is fixed if the As64BitReg() is removed. When this is done, LoadBaseIndexed() sees that the index is a 32-bit integer and introduces a zero-extend. This is currently done with a separate instruction. It was on my plans to improve this an do the sign extension and the load in one single instruction. I postponed working on this because I estimated this was not very beneficial anyway, but if the As64BitReg() is removed for aget and aput, things may be suddenly different.

Patch-set: 1