Update patch set 17
Abandoned

So we need to calculate BB information because pass runs so early before it is done for first time. Trying to do a subset of the work is a bit hacky because there are dependencies. For example, this pass needs topological sort, which needs dominators, which needs ssa, etc. So because of the increase in compilation time I will abandon this. Thanks for all the reviewers that helped ensure its correctness. :)

Patch-set: 17
Status: abandoned