Update patch set 4
Patch Set 4:

Let me give some thought on how a test for this might look like. Though I imagine it might be hard to devise a test since some of the infrastructure is lacking. We use use-def chains to find users of a define ( https://android-review.googlesource.com/#/c/91465/ ). UD chains cannot be complete with missing ssa data. Not being complete makes it seem that if something has no uses, it might be dead when in reality it is not.

Patch-set: 4