Update patch set 1
Patch Set 1: Code-Review-1

Spaghetti code in the middle again. You take half of the fixes from the WIP verbatim and incorporate the rest, but then don't do the refactoring of taking Heap and Dalvik regs apart to make it readable, *without* an explanation and *any* code comments in the critical part. E.g., where's the explanation that this is conservative? I think you tried to incorporate as part of the conjunction (though I'm not sure it's right), but for a person starting to read this code, this is horrible (i.e., as it was before).

Also, no tests. LSE is/was broken and that was not detected before. I am not comfortable of merging this without some corner case testing.

Patch-set: 1
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1