Update patch set 4
Patch Set 4:

Bill,
So I can do the trivial test which is to check a CFG with all different blocks whether they have ssa for every instruction. Obviously this is a bit too conservative because this is not a test of the real problem. Do YOU have any comments on what I should test?

Another approach to deal with this problem is to factor out this recalculation functionality into an overridable method. For AOSP we leave the logic same. What do you think?

Patch-set: 4