Update patch set 1
Patch Set 1:

> Could you be more precise? What naming conflicts? And why have this
 > ginormous change, if the change in the optimizing compiler would
 > likely be smaller?

I understand this touches many lines for a renaming, but I believe it is worth it. Two reasons for this patch:
1) It improves the current naming by having clearly prefixed register enums 'X', 'W', 'D', 'S'. (With the current scheme, what is a 'Register'? Is a 'Wregister', 'Dregister', or 'SRegister' a 'Register'?
2) It will make the opt compiler code much easier to write and more readable.
The opt compiler code generation will rely heavily on VIXL. To avoid having to prefix everywhere with 'vixl::', we will import the vixl namespace in -arm64.cc files (only -  to not pollute other archs).
VIXL exposes the 'Register' type (covering both X and W registers), which is used everywhere in the VIXL helpers and will be used everywhere in the codegen.

Patch-set: 1