Update patch set 1
Patch Set 1:

> Any reason for switching SP? For the ARM-proposed renaming?

Not sure what you are asking.
We want XZR/WZR to use the same code 31 to be coherent and match VIXL. Then we need SP to use a different code from the ZR registers to be able to differentiate them. VIXL uses the code 64, but we can't use it hear to keep the enum contiguous.

Patch-set: 1