Update patch set 2
Patch Set 2:

This is just wrong. Consider a loop where the register in null-checked at entry and becomes null later. On the first pass we'll mark the insn as null-checked and then never unmark it. (The lack of tests that would catch this is really bad.)

The correct approach is to wait until we converge and then do a final pass as we do for the GVN.

Patch-set: 2
Reviewer: Vladimir Marko <1018108@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0