Update patch set 2
Patch Set 2:

Vlad,

I am eager to see your unit test for this because I admit I fail to understand the scenario you describe. From what I gather, it seems more to me that what you are describing is an assumption made in the NCE algorithm.

This is the scenario that I believe you are talking about:

while (i < 10) {
  a.field = 1;
  a = null;
  a.field = 2;
  i++;
}

The NCE algorithm should see that the redefinition of "a" is a different object that also needs null checked.

Anyway, I think a test case should be enough for me to gather where the problem comes from since I am surprised that this patch caused it :)

Patch-set: 2