Update patch set 1
Patch Set 1:

I explained to Ian before but motivation of this patch is to allow optimizations like LICM and DCE. So what this patch does is that currently it leaves the reference map calculator disabled. If any pass actually requests for this, only then it will figure out which regs (including which compiler temps) actually contain references.

Patch-set: 1
Label: Verified=0