Update patch set 3
Patch Set 3:

I plan to get it merged ASAP, there will be conflicts in particular with:
https://android-review.googlesource.com/#/c/108630/
as this change did multiple refactorings in one go and the change above is the better clean-up. I wouldn't bother rebasing in the short-term.
I still have changes I want to this CL and plan to look at it after 108630 goes in.

Patch-set: 3