Update patch set 2
Patch Set 2: Verified-1

This is an automated message from the Deckard Autoverifier.

Patchset merges but does not build in AOSP.
The following snippet of `make` output may point you in the right direction:

  In file included from art/compiler/optimizing/builder.cc:18:
In file included from art/compiler/optimizing/builder.h:27:
[1mart/compiler/optimizing/nodes.h:1696:9: [0m[0;1;31merror: [0m[1m'next_' is a private member of 'art::HInstruction'[0m
    if (next_ == nullptr
[0;1;32m        ^
[0m[1mart/compiler/optimizing/nodes.h:761:17: [0m[0;1;30mnote: [0mdeclared private here[0m
  HInstruction* next_;
[0;1;32m                ^
[0m[1mart/compiler/optimizing/nodes.h:1697:13: [0m[0;1;31merror: [0m[1m'next_' is a private member of 'art::HInstruction'[0m
...


If your patchset has a dependency in a separate project, please ensure that
both changes are uploaded with the same Change-Id.

Patch-set: 2
Label: Verified=-1