Update patch set 2
Patch Set 2: Code-Review-1

build/core/clang/config.mk has CLANG_CONFIG_UNKNOWN_CFLAGS, but that won't help here. The issue is that they are only preprocessing the files (i.e. passing -E on the command line), but then also passing an argument that is only used by the assembler (which won't be invoked at all). I think that it is more of a bug on the ccache side of things than anything else. If they are going to mutate flags, they need to do a better job of converting them to something that makes sense.

Patch-set: 2
Reviewer: Stephen Hines <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1