Update patch set 2
Patch Set 2:

The --noexecstack is getting passed to the assembler, which may or may not be the integrated assembler, so I have no idea if there is a way to only insert it on ccache's second pass. I think that ultimately ccache is not something that we want to encourage use of, since it is not equivalent to compiling the original file (based on that link Ying provided and its discussion).

Patch-set: 2