Update patch set 2
Patch Set 2:

It's a ccache bug because ccache is doing something very nonstandard here (by design). It is intercepting a compilation call and then altering the flags (i.e. adding "-E"). When it does this, feel free to have ccache also remove any flags that are invalid/incorrect for a preprocessor-only run. Since we don't use ccache internally, we have no motivation to change our behavior. Making this just a warning (with no error) might cause us to miss an important future error (namely adding what we thought was a valid flag that is actually doing nothing in practice).

Patch-set: 2