Update patch set 2
Patch Set 2: Code-Review-1

Ah I see, thanks for the additional detail.

For our part, we use ccache for developer builds as it saves a great deal of time (orders of magnitude faster builds), but never use it when cutting actual releases.

Looks like the way to move forward is to try to fix this in build/ as Ying suggests.

Patch-set: 2
Reviewer: Andrew P Boie <1011333@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1