Update patch set 3
Patch Set 3:

Rebased.

This workaround patch continues to be restricted to just Mips, as Ian requested.   Stephan's -1 review said that doing a Mips-specific workaround is unsafe, in principle.  And a source change to the casting methods could eliminate the compilers' confusions without costing performance on Arm.

The Mips compiler team has discovered the degraded optimization between 4.8 and 4.9 which uncovered this latent issue.  What changed, was a adjusted heuristic limit for inline expansions.

Patch-set: 3
Label: Verified=0