Update patch set 1
Patch Set 1:

This script is quite useful for ART developers to debug the code generated by ART compiler at the moment. Currently the oat file is not opened by dlopen(), and in gdb-gdbserver environment, gdb-jit also doesn't work. The only way make gdb understand JAVA symbols is to load the symbols from oat file to its mapped address, which is quite inconvenient. This script helps us to this.

And it takes the assumption that we always have the symbol information in oat files, which might not be true some day if we change the behavior of dex2oat, build script or run-test script. I don't expect this script can not be broken and always be useful. It can be even useless once we move to dlopen().

Patch-set: 1