Update patch set 1
Patch Set 1:

The motivation for this change is when you run tests with libart.so you don't want to compile the debug versions as well.

Unfortunately, this CL is not quite working yet as 1) some tests explicitly ask for both and 2) the makefile test dependencies depend on both versions.

Patch-set: 1
Label: Verified=0