Update patch set 1
Patch Set 1:

I understand that it is useful as is, because people mostly test the run-tests in their default config, which is to use dex2oatd, which then puts the symbol information in. However, ngeoffray is just adding the wiring to correctly use the non-debug versions, also, which does *not* have symbols.

So I feel it's a simple step to add the symbolizer in, and the tool would remain valid.

Patch-set: 1
Reviewer: Andreas Gampe <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0