Update patch set 1
Patch Set 1:

Vladimir, this approach has been suggested in https://android-review.googlesource.com/#/c/107261/. Could you please read that thread?
That patch contains the alternative solution to this problem but it was considered as too complex. That why it was proposed to resolve it in this way.

Patch-set: 1
Reviewer: Serguei I Katkov <1040038@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0