Update patch set 1
Patch Set 1:

Yes, x86 needs more thought. But on arm64 it's pretty simple to use the 32-bit index and avoid the issue altogether.

Note that LoadBaseIndexed() actually avoids the issue (see also the TODO there), so the bug in arm64 GenArrayGet()/GenArrayPut() is currently limited to the wide/fp case where we calculate the address instead of using indexed load/store. And that address calculation should not be performed at all as the arm64 has indexed load/store for wide/fp regs. It's time to move away from the straight-forward arm port.

Patch-set: 1