Update patch set 1
Patch Set 1:

Andreas, you are correct. It's mainly for some crash issues in MTBF testing and can't be debugged online.

I know we could use command line to enable the debugging features. But it means we need to change the zygote start command line (we need to create core dump for zygote itself sometime) which is not so runtime...

Patch-set: 1
Reviewer: Fengwei Yin <1011068@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0