Update patch set 2
Patch Set 2: Code-Review-2

(2 comments)

There's a bunch of work in this change not relevant to the bug. The actual bug appears to be re-entrant calling of the dumping code by suspending in the verifier. We should just avoid/ignore this checkpoint once we're processing it.

Patch-set: 2
Reviewer: Ian Rogers <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2