Update patch set 1
Patch Set 1:

As I understand it, 0001 and 0004's v0 are the same RegLocation, because they're the same SSA reg definition, v0_0. If you don't set .fp because of the use in 0001, then .fp won't be set for its use in 0004, when it needs to be.

Patch-set: 1
Reviewer: Stephen Kyle <1053354@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0