Update patch set 1
Patch Set 1:

Some historical info:  In the very early (interpreter-only) days of Dalvik, reducing dex file size was considered extraordinarily important.  In that light, treating literals as bit patterns and reusing when possible was intentional and early versions of dx went out of their way to share literals.   We recognized that this was a poor choice for compiled code early in Art development, and dx was changed to not do this.

However, code produced by ancient dx versions still exists in the wild - though I at this point I believe it to be very, very rare.  We do have to handle it, but skipping compilation when encountering it is not unreasonable.

That said, changing InToRegStorageMapper to take shorties rather than RegLocations seems like the right thing to do.

Patch-set: 1
Reviewer: Bill Buzbee <1001578@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0