Update patch set 2
Patch Set 2: Verified+1

The test succeeds with this change, without it it fails the field ordering check but (if that one is fixed) not the reference offset check. I don't know how to reproduce the latter, so I'm going to submit this as is.

Patch-set: 2
Reviewer: Vladimir Marko <1018108@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1