Update patch set 1
Patch Set 1:

So the test does not make sense for prebuild. The run script removes it from the flags. I think we did it this way because of our ongoing test infrastructure work not being complete. Instead of working around it in run scripts, we should change Android.run-test.mk to black-list the test (which is what your change to the makefile is touching with).

Patch-set: 1