Update patch set 1
Patch Set 1:

I think we should amend this change. I'm in discussion with frameworks people to establish a PIC build. The issue is that this is basically moving everything to an -eng build, which is too slow. PIC would solve this as we would not have direct pointers.

Could you take a look at the PIC changes and make the full deopt dependent on whether the boot image is PIC or not?

Patch-set: 1
Label: Code-Review=0