Update patch set 1
Patch Set 1:

For a Java add method like "int f(int x, int y) { return x + y; }", without this patch we get:

  sub sp, sp, #0x30 (48)
  stp x20, x21, [sp, #16]
  stp x22, x30, [sp, #32]
  str w0, [sp]
  mov w21, w1
  mov w22, w2
  add w20, w21, w22
  mov w0, w20
  ldp x20, x21, [sp, #16]
  ldp x22, x30, [sp, #32]
  add sp, sp, #0x30 (48)
  ret

With the patch we get:

  sub sp, sp, #0x10 (16)
  add w3, w1, w2
  mov w0, w3
  add sp, sp, #0x10 (16)
  ret
 
Note also that the optimisation applies to ~4% of the methods in boot.oat so far. Some of them are less trivial than a simple add. These include hash functions, bit manipulation functions, etc. I suspect this percentage can be increased (as leaf detection is not optimal right now). The hard upper limit is ~14% (this is the percentage of DEX leaf methods in boot.oat).

Outstanding things to do include:

1. Support kLeafOptimisation == false. This involves allowing to call suspend checks handlers. We should also support explicit stack checks.
2. Make sure the optimisation can be disabled/enabled similarly to all the other optimisations.
3. Make leaf detection optimal (e.g. IF, GOTO opcodes currently make the method non-leaf, while - in principle - this should happen only when branching backward).

One of the main problems I have with the leaf-detection patch is that it requires consistency between the prediction phase (where we predict whether a MIR opcode will generate a helper call) and the generation phase (where the calls are actually generated). Currently, I catch inconsistencies and abort. Ideally, the prediction functionality can be fully derived from the generation functionality. In other words, given a function that generates the code for a given MIR opcode, including calls to helpers, it is possible to derive a function that predicts whether a MIR opcode will generate calls to helpers. In practice, this derivation is difficult to do in C++ and I see that even in the optimizing compiler two distinct functions are provided which need to be consistent (e.g. LocationBuilderARM vs InstructionCodeGeneratorARM).

Patch-set: 1
Reviewer: Matteo Franchin <1037505@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0